requirements.txt
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
attrs==25.4.0
Authlib==1.6.7
backports.tarfile==1.2.0
beartype==0.22.9
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
croniter==6.0.0
cryptography==46.0.4
cyclopts==4.5.2
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fakeredis==2.33.0
fastapi==0.128.6
fastmcp==2.14.5
filetype==1.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-genai==1.62.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.2.9
langchain-core==1.2.9
langchain-google-genai==4.2.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
lupa==2.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
narwhals==2.16.0
numpy==2.2.6
openapi-pydantic==0.5.1
opencv-python==4.13.0.92
opentelemetry-api==1.39.1
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
pathable==0.4.4
pathvalidate==3.3.1
pdf2image==1.17.0
pillow==12.1.0
platformdirs==4.5.1
prometheus_client==0.24.1
protobuf==6.33.5
py-key-value-aio==0.3.0
py-key-value-shared==0.3.0
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pydocket==0.17.7
Pygments==2.19.2
PyJWT==2.11.0
PyMuPDF==1.26.7
pypdf==6.7.0
pyperclip==1.11.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
PyYAML==6.0.3
redis==7.1.1
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rich-rst==1.3.2
rpds-py==0.30.0
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sse-starlette==3.2.0
starlette==0.52.1
taskgroup==0.2.2
tenacity==9.1.4
toml==0.10.2
tomli==2.4.0
tornado==6.5.4
typer==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uv==0.10.2
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
